Add support for NaN comparisons
Checks before every single conditional jump or set to make sure that not
all of ZF, CF and PF are set, indicating an unordered result. If this
result is found, and additional dummy comparison is made which evaluates
to false for whatever the conditional jump was checking for (except !=,
which is always true).

Also had to rewrite the do-while logic so that it does "if condition is
zero break" instead of "if condition is non-zero loop again". This is
more consistent with the other loops, and enables this kind of loop to
also break when the condition is unordered (as is required)