Add pointers support
Also involved rewriting how compound operators (+=, etc.) are treated
because of a new requirement where if you have an expression such as
```C
*print_hi_and_return_5() += 3;
```
then the expression on the left must be evaluated *exactly once* when
this line is run. This means that the previous method I was using, which
rewrote the expression as `*something = *something + 5` was an absolute
nonstarter. I introduced a new Compound expression type and evaluated
the entire expression (including the 2 required type-casting steps for
expressions with mixed types) in the IR package.