Add parsing and validation for struct definitions
This implementation is a lot closer (though to my knowledge also not
100% conformant yet) with the C standard for struct definitions compared
to what's outlined in the book. It allows for struct declarations and
definitions in the following forms
```c
struct a;
```
```c
struct a v;
```
```c
struct a v = {1};
```
```c
struct a {
    int b;
};
```
```c
struct a {
    int b
} v;
```
```c
struct a {
    int b;
} v = {1};
```
```c
struct c {
    d int;
};
struct a {
    // struct c already defined in scope so no need to define it here
    struct c nested_c;
    struct e {
f int;
    } nested_d;
} v = {{1}, {1}}
```
etc.