Finish implementing unions
Unions add a pretty significant wrinkle to the existing classify_struct
function which determines which register a struct parameter or return
value should occupy, since for example, a union of [Int, Double[2]]
would require the specific classes [Integer, SSE] because the int value
can't be stored in SSE, but it is not big enough to override the second
class. Debugging that took quite a while because I didn't realise part
of the test cases involved compiling one file with `gcc` and checking
that it inter-operates properly with code outputted from my compiler,
which caused some headaches because I wasn't able to reproduce the
failure with my testing setup by just compiling both files with my
compiler.