Implement Register Coalescing, completing book content
Thus completing a compiler which implements every feature described in
Writing a C Compiler, and testing it against the provided test suite. I
still want to add a few more things (and probably a README) before I'm
happy to call this project done, but this in itself is a major
milestone.