Add (hacky, temporary) support for anoymous function params and `const`
Right now that just means const is silently ignored and anonymous params
are renamed to some temporary name and (notably) function definitions
(ie with bodies) are not validated to make sure they have no anonymous
params. This was done in an effort to see if I could get math.h to
compile with this compiler, and while a lot of progress was made during
this experiment, I realised that I'm going to need to implement my own
preprocessor if I want to make any further headway on this project.