Implement preprocessor first pass and lexer
These compose the first 2 (technically 3 except step 2 of 3 is "pretend
trigraphs do not exist and cry otherwise") steps of the c preprocessor,
following this PDF which has been incredibly helpful for exactly this:
https://gcc.gnu.org/onlinedocs/cpp.pdf

Now comes the fun part, the actual interpreter and defining a set of
default definitions to come with this compiler/preprocessor pair. This
approach should make interoperability with the C standard library a much
more achievable goal.