Add enums
This is a little more spaghetti code-ish than usual because of how enums
interact with scopes, and the fact that my type-checked generally has no
knowledge of scopes (scopes are handled in the parsing stage to break
ambiguity involving `typedef`).

Because of this, I needed to make sure that most any code from enums
that would result in additional values being brought into scope is
handled immediately in the parsing stage, rather than the type checking
stage (which is where, say, struct and union members are resolved at the
moment).