Fix 2 issues with var_args, add octal parsing
Namely, %al is meant to store the number of XMM args passed to a
variadic function when the function is called (ie this is set by the
caller). I had *literally no idea* this was the case and was baffled at
how printf seemed to just randomly fail all the time.

Also fix float promotion to double, another requirement for System V.

184/220