Last round of fixes for c-testsuite for now (188/220)
Also add the helper files that I used in the c-testsuite folder locally
for testing, so that these tests can be run with the same skip list
later.

Includes several improvements to the preprocessor and how types are
handled, and removes the requirement for a constant variable to be
initialised. Any further improvements would require implementing a major
feature (eg variable-length arrays), or would require restructuring the
code a lot to handle edge cases (such as expressions of the form `a++,
b++`) which are sometimes seen in `for` loops.